Fix UI Scaling bug
Players have complained that sometimes the buttons to close a mod's
config (i.e., Cancel or "Save & Close") are outside the screen so they
can't click it. Pressing "Esc" or "E" also doesn't close the dialog.

It is because GMCM is still using Game1.viewport which has the gameworld
coordinates, which on players using gameworld zoom < UI zoom will cause
the coordinates of the menu elements to be outside the UI viewport,
hence not seen.

I brute-fixed this by replacing all instances of Game1.viewport with
Game1.uiViewport, and at the very least this bug is no fixed.

Also, I added some "null catching" code in RegisterModMenu() because
there had been spurious reports of GMCM not working well with
CustomMusic (something I cannot replicate on my machine)

~pepoluan